[Refactor] Refactor viewmodel factory with AuthRepository
