+++ app/src/main/java/org/sopt/sample/presentation/login/ViewModel/LoginViewModel.kt
@@ -1,0 +1,9 @@
+package org.sopt.sample.presentation.login.ViewModel
+
+import androidx.lifecycle.ViewModel
+import org.sopt.sample.domain.repository.AuthRepository
+
+class LoginViewModel(private val authRepository: AuthRepository) :ViewModel() {
+
+
+}
