--- app/src/main/java/org/sopt/sample/presentation/signup/viewmodel/SignUpViewModelFactory.kt
@@ -1,17 +1,0 @@
-package org.sopt.sample.presentation.signup.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import org.sopt.sample.domain.repository.AuthRepository
-
-@Suppress("UNCHECKED_CAST")
-class SignUpViewModelFactory(
-    private val authRepository: AuthRepository
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(SignUpViewModel::class.java)) {
-            return SignUpViewModel(authRepository) as T
-        }
-        throw IllegalArgumentException("ViewModel class not found")
-    }
-}
